feat: improve minimap navigation and runtime stability

+++ app/src/main/java/org/luaj/LuaString.java
@@ -96,7 +96,7 @@ public class LuaString extends LuaValue {
             }
             value = (value * i) + digit;
         }
-        return Long.valueOf(negative ? -value : value);
+        return negative ? -value : value;
     }
 
     private boolean a(byte[] bArr, int i, int i2) {
@@ -183,7 +183,7 @@ public class LuaString extends LuaValue {
             if (digit < 0 || digit >= base) return null;
             value = value * base + digit;
         }
-        return Long.valueOf(negative ? -value : value);
+        return negative ? -value : value;
     }
 
     public static int encodeToUtf8(char[] cArr, int i, byte[] bArr, int i2) {
@@ -943,12 +943,12 @@ public class LuaString extends LuaValue {
 
     public LuaValue tointeger() {
         Long e = e();
-        return e == null ? LuaValue.NIL : LuaValue.valueOf(e.longValue());
+        return e == null ? LuaValue.NIL : LuaValue.valueOf(e);
     }
 
     public LuaValue tointeger(int i) {
         Long a = a(i);
-        return a == null ? LuaValue.NIL : LuaValue.valueOf(a.longValue());
+        return a == null ? LuaValue.NIL : LuaValue.valueOf(a);
     }
 
     public String tojstring() {
